test: keep image fixtures off photon
