+++ src/plugin-sdk/test-helpers/image-fixtures.ts
@@ -1,0 +1,129 @@
+import { deflateSync } from "node:zlib";
+import { encodePngRgb, encodePngRgba } from "../../media/png-encode.js";
+
+type Rgba = {
+  r: number;
+  g: number;
+  b: number;
+  a?: number;
+};
+
+const PNG_SIGNATURE = Buffer.from([0x89, 0x50, 0x4e, 0x47, 0x0d, 0x0a, 0x1a, 0x0a]);
+const TINY_JPEG = Buffer.from(
+  "/9j/4AAQSkZJRgABAQAAAQABAAD/2wBDAP//////////////////////////////////////////////////////////////////////////////////////2wBDAf//////////////////////////////////////////////////////////////////////////////////////wAARCAABAAEDASIAAhEBAxEB/8QAFQABAQAAAAAAAAAAAAAAAAAAAAX/xAAUEAEAAAAAAAAAAAAAAAAAAAAA/9oADAMBAAIQAxAAAAH/xAAUEAEAAAAAAAAAAAAAAAAAAAAA/9oACAEBAAEFAqf/xAAUEQEAAAAAAAAAAAAAAAAAAAAA/9oACAEDAQE/ASP/xAAUEQEAAAAAAAAAAAAAAAAAAAAA/9oACAECAQE/ASP/xAAUEAEAAAAAAAAAAAAAAAAAAAAA/9oACAEBAAY/Aqf/xAAUEAEAAAAAAAAAAAAAAAAAAAAA/9oACAEBAAE/IV//2gAMAwEAAgADAAAAEP/EFBQRAQAAAAAAAAAAAAAAAAAAABD/2gAIAQMBAT8QH//EFBQRAQAAAAAAAAAAAAAAAAAAABD/2gAIAQIBAT8QH//EFBABAQAAAAAAAAAAAAAAAAAAABD/2gAIAQEAAT8QH//Z",
+  "base64",
+);
+const CRC_TABLE = (() => {
+  const table = new Uint32Array(256);
+  for (let index = 0; index < table.length; index += 1) {
+    let value = index;
+    for (let bit = 0; bit < 8; bit += 1) {
+      value = value & 1 ? 0xedb88320 ^ (value >>> 1) : value >>> 1;
+    }
+    table[index] = value >>> 0;
+  }
+  return table;
+})();
+
+function crc32(buffer: Buffer): number {
+  let crc = 0xffffffff;
+  for (const byte of buffer) {
+    crc = CRC_TABLE[(crc ^ byte) & 0xff] ^ (crc >>> 8);
+  }
+  return (crc ^ 0xffffffff) >>> 0;
+}
+
+function pngChunk(type: string, data: Buffer): Buffer {
+  const typeBuffer = Buffer.from(type, "ascii");
+  const length = Buffer.alloc(4);
+  length.writeUInt32BE(data.length, 0);
+  const crc = Buffer.alloc(4);
+  crc.writeUInt32BE(crc32(Buffer.concat([typeBuffer, data])), 0);
+  return Buffer.concat([length, typeBuffer, data, crc]);
+}
+
+function fillSolidRgba(width: number, height: number, color: Rgba): Buffer {
+  const pixels = Buffer.alloc(width * height * 4);
+  for (let offset = 0; offset < pixels.length; offset += 4) {
+    pixels[offset] = color.r;
+    pixels[offset + 1] = color.g;
+    pixels[offset + 2] = color.b;
+    pixels[offset + 3] = color.a ?? 255;
+  }
+  return pixels;
+}
+
+function fillSolidRgb(width: number, height: number, color: Rgba): Buffer {
+  const pixels = Buffer.alloc(width * height * 3);
+  for (let offset = 0; offset < pixels.length; offset += 3) {
+    pixels[offset] = color.r;
+    pixels[offset + 1] = color.g;
+    pixels[offset + 2] = color.b;
+  }
+  return pixels;
+}
+
+export function createSolidPngBuffer(width: number, height: number, color: Rgba): Buffer {
+  if (color.a === undefined || color.a === 255) {
+    return encodePngRgb(fillSolidRgb(width, height, color), width, height);
+  }
+  return encodePngRgba(fillSolidRgba(width, height, color), width, height);
+}
+
+export function createNoisyPngBuffer(width: number, height: number): Buffer {
+  const rgba = createNoisyRgbaBuffer(width, height);
+  const rgb = Buffer.alloc(width * height * 3);
+  for (let source = 0, target = 0; source < rgba.length; source += 4, target += 3) {
+    rgb[target] = rgba[source] ?? 0;
+    rgb[target + 1] = rgba[source + 1] ?? 0;
+    rgb[target + 2] = rgba[source + 2] ?? 0;
+  }
+  return encodePngRgb(rgb, width, height);
+}
+
+export function createGrayscaleAlphaPngBuffer(width: number, height: number): Buffer {
+  const stride = width * 2;
+  const raw = Buffer.alloc((stride + 1) * height);
+  for (let row = 0; row < height; row += 1) {
+    const rawOffset = row * (stride + 1);
+    raw[rawOffset] = 0;
+    for (let column = 0; column < width; column += 1) {
+      const pixel = rawOffset + 1 + column * 2;
+      const seed = row * width + column;
+      raw[pixel] = seed % 256;
+      raw[pixel + 1] = seed % 5 === 0 ? 96 : 255;
+    }
+  }
+
+  const ihdr = Buffer.alloc(13);
+  ihdr.writeUInt32BE(width, 0);
+  ihdr.writeUInt32BE(height, 4);
+  ihdr[8] = 8;
+  ihdr[9] = 4;
+  ihdr[10] = 0;
+  ihdr[11] = 0;
+  ihdr[12] = 0;
+
+  return Buffer.concat([
+    PNG_SIGNATURE,
+    pngChunk("IHDR", ihdr),
+    pngChunk("IDAT", deflateSync(raw)),
+    pngChunk("IEND", Buffer.alloc(0)),
+  ]);
+}
+
+export function createTinyJpegBuffer(): Buffer {
+  return Buffer.from(TINY_JPEG);
+}
+
+export function createNoisyRgbaBuffer(width: number, height: number): Buffer {
+  const pixels = Buffer.alloc(width * height * 4);
+  for (let offset = 0; offset < pixels.length; offset += 4) {
+    const seed = offset / 4;
+    pixels[offset] = seed % 251;
+    pixels[offset + 1] = (seed * 17) % 253;
+    pixels[offset + 2] = (seed * 29) % 255;
+    pixels[offset + 3] = 255;
+  }
+  return pixels;
+}
